# perf(fetch): improve body mixin methods