--- a/lib/web/fetch/body.js
+++ b/lib/web/fetch/body.js
@@ -318,7 +318,8 @@ function bodyMixinMethods (instance) {
       // given a byte sequence bytes: return a new ArrayBuffer
       // whose contents are bytes.
       return consumeBody(this, (bytes) => {
-        return new Uint8Array(bytes).buffer
+        // Note: arrayBuffer already cloned.
+        return bytes.buffer
       }, instance)
     },
 
@@ -432,7 +433,7 @@ async function consumeBody (object, convertBytesToJSValue, instance) {
   // 5. If object’s body is null, then run successSteps with an
   //    empty byte sequence.
   if (object[kState].body == null) {
-    successSteps(new Uint8Array())
+    successSteps(Buffer.allocUnsafe(0))
     return promise.promise
   }